Removed wrapper from model_provider.py
